sitemap/dynamic-urls: dynamic urls fixes
