+++ src/sitemap/dynamic_urls.rs
@@ -1,3 +1,5 @@
+use crate::sitemap::section::{Section, Subsection};
+
 #[derive(Debug, serde::Deserialize, Clone)]
 pub struct DynamicUrlsTemp {
     #[serde(rename = "dynamic-urls-body")]
@@ -48,7 +50,53 @@ impl DynamicUrls {
 
     // If any one does not have path parameters so return true
     pub fn not_have_path_params(&self) -> bool {
-        fpm::sitemap::section::Section::not_contains_path_params(&self.sections)
+        fn check_toc(toc: &fpm::sitemap::toc::TocItem) -> bool {
+            if toc.path_parameters.is_empty() {
+                return true;
+            }
+
+            for toc in toc.children.iter() {
+                if check_toc(toc) {
+                    return true;
+                }
+            }
+            false
+        }
+
+        fn check_sub_section(sub_section: &Subsection) -> bool {
+            // Note: No need check subsection
+            // if sub_section.path_parameters.is_empty() {
+            //     return true;
+            // }
+
+            for toc in sub_section.toc.iter() {
+                if check_toc(toc) {
+                    return true;
+                }
+            }
+            false
+        }
+
+        fn check_section(section: &Section) -> bool {
+            // Note: No need check section
+            // if section.path_parameters.is_empty() {
+            //     return true;
+            // }
+
+            for sub_section in section.subsections.iter() {
+                if check_sub_section(sub_section) {
+                    return true;
+                }
+            }
+            false
+        }
+
+        for section in self.sections.iter() {
+            if check_section(section) {
+                return true;
+            }
+        }
+        false
     }
 }
